services/jupyter/requirements.txt
requests
weaviate-client
pdf2image